Move build utils to appropriate file
