+++ docker/utils/__init__.py
@@ -1,13 +1,13 @@
 # flake8: noqa
-from .build import tar, exclude_paths
+from .build import create_archive, exclude_paths, mkbuildcontext, tar
 from .decorators import check_resource, minimum_version, update_headers
 from .utils import (
     compare_version, convert_port_bindings, convert_volume_binds,
-    mkbuildcontext, parse_repository_tag, parse_host,
+    parse_repository_tag, parse_host,
     kwargs_from_env, convert_filters, datetime_to_timestamp,
     create_host_config, parse_bytes, parse_env_file, version_lt,
     version_gte, decode_json_header, split_command, create_ipam_config,
     create_ipam_pool, parse_devices, normalize_links, convert_service_networks,
-    format_environment, create_archive, format_extra_hosts
+    format_environment, format_extra_hosts
 )
